+++ pybop/optimisers/pints_optimisers.py
@@ -1,10 +1,7 @@
-from warnings import warn
-
 from pints import CMAES as PintsCMAES
 from pints import PSO as PintsPSO
 from pints import SNES as PintsSNES
 from pints import XNES as PintsXNES
-from pints import Adam as PintsAdam
 from pints import IRPropMin as PintsIRPropMin
 from pints import NelderMead as PintsNelderMead
 
@@ -78,83 +75,6 @@ class GradientDescent(BasePintsOptimiser):
         super().__init__(
             cost,
             GradientDescentImpl,
-            max_iterations,
-            min_iterations,
-            max_unchanged_iterations,
-            multistart,
-            parallel,
-            **optimiser_kwargs,
-        )
-
-
-class Adam(BasePintsOptimiser):
-    """
-    Implements the Adam optimisation algorithm.
-
-    This class extends the Adam optimiser from the PINTS library, which combines
-    ideas from RMSProp and Stochastic Gradient Descent with momentum.
-
-    Note that this optimiser does not support boundary constraints.
-
-    Parameters
-    ----------
-    cost : callable
-        The cost function to be minimized.
-    max_iterations : int, optional
-        Maximum number of iterations for the optimisation.
-    min_iterations : int, optional (default=2)
-        Minimum number of iterations before termination.
-    max_unchanged_iterations : int, optional (default=15)
-        Maximum number of iterations without improvement before termination.
-    multistart : int, optional (default=1)
-        Number of optimiser restarts from randomly sample position. These positions
-        are sampled from the priors.
-    parallel : bool, optional (default=False)
-        Whether to run the optimisation in parallel.
-    **optimiser_kwargs : optional
-        Valid PINTS option keys and their values, for example:
-        x0 : array_like
-            Initial position from which optimisation will start.
-        sigma0 : float
-            Initial step size or standard deviation depending on the optimiser.
-        bounds : dict
-            A dictionary with 'lower' and 'upper' keys containing arrays for lower and
-            upper bounds on the parameters.
-        use_f_guessed : bool
-            Whether to return the guessed function values.
-        absolute_tolerance : float
-            Absolute tolerance for convergence checking.
-        relative_tolerance : float
-            Relative tolerance for convergence checking.
-        max_evaluations : int
-            Maximum number of function evaluations.
-        threshold : float
-            Threshold value for early termination.
-
-    See Also
-    --------
-    pints.Adam : The PINTS implementation this class is based on.
-    """
-
-    warn(
-        "Adam is deprecated and will be removed in a future release. Please use AdamW instead.",
-        DeprecationWarning,
-        stacklevel=2,
-    )
-
-    def __init__(
-        self,
-        cost,
-        max_iterations: int = None,
-        min_iterations: int = 2,
-        max_unchanged_iterations: int = 15,
-        multistart: int = 1,
-        parallel: bool = False,
-        **optimiser_kwargs,
-    ):
-        super().__init__(
-            cost,
-            PintsAdam,
             max_iterations,
             min_iterations,
             max_unchanged_iterations,