+++ tests/unit/test_observers.py
@@ -1,8 +1,8 @@
 import numpy as np
+import pybamm
 import pytest
 
 import pybop
-from examples.standalone.model import ExponentialDecay
 
 
 class TestObserver:
@@ -27,15 +27,17 @@ class TestObserver:
 
     @pytest.fixture(params=[1, 2])
     def model(self, parameters, request):
-        parameter_set = {"k": 0.1, "y0": 1.0}
-        model = ExponentialDecay(parameter_set=parameter_set, n_states=request.param)
+        parameter_set = pybamm.ParameterValues({"k": 0.1, "y0": 1.0})
+        model = pybop.ExponentialDecayModel(
+            parameter_set=parameter_set, n_states=request.param
+        )
         model.build(parameters=parameters)
         return model
 
     @pytest.mark.unit
     def test_observer(self, model, parameters):
         n = model.n_states
-        observer = pybop.Observer(parameters, model, signal=["2y"])
+        observer = pybop.Observer(parameters, model, signal=["y_0"])
         t_eval = np.linspace(0, 20, 10)
         expected = parameters["y0"].initial_value * np.exp(
             -parameters["k"].initial_value * t_eval
@@ -49,7 +51,7 @@ class TestObserver:
             )
             np.testing.assert_array_almost_equal(
                 observer.get_current_measure(),
-                np.array([[2 * y]]),
+                np.array([[y]]),
                 decimal=4,
             )
 
@@ -60,7 +62,7 @@ class TestObserver:
             ValueError, match="values and times must have the same length."
         ):
             observer.log_likelihood(
-                {"2y": t_eval}, np.array([1]), inputs=observer._state.inputs
+                {"y_0": t_eval}, np.array([1]), inputs=observer._state.inputs
             )
 
         # Test covariance
@@ -79,12 +81,12 @@ class TestObserver:
                 "Output": expected,
             }
         )
-        observer._target = {"2y": expected}
+        observer._target = {"y_0": expected}
         observer.evaluate(parameters.as_dict())
 
     @pytest.mark.unit
     def test_unbuilt_model(self, parameters):
-        model = ExponentialDecay()
+        model = pybop.ExponentialDecayModel()
         with pytest.raises(
             ValueError, match="Only built models can be used in Observers"
         ):
@@ -94,10 +96,13 @@ class TestObserver:
     def test_observer_inputs(self):
         initial_state = {"Initial open-circuit voltage [V]": 4.0}
         t_eval = np.linspace(0, 1, 100)
-        model = ExponentialDecay(n_states=1)
+        model = pybop.ExponentialDecayModel(n_states=1)
         model.build()
         observer = pybop.Observer(
-            pybop.Parameters(), model, signal=["y_0", "2y"], initial_state=initial_state
+            pybop.Parameters(),
+            model,
+            signal=["y_0", "y_0"],
+            initial_state=initial_state,
         )
         assert observer.initial_state == initial_state
 
@@ -106,5 +111,5 @@ class TestObserver:
             match="Observer.log_likelihood is currently restricted to single output models.",
         ):
             observer.log_likelihood(
-                {"2y": t_eval}, t_eval, inputs=observer._state.inputs
+                {"y_0": t_eval}, t_eval, inputs=observer._state.inputs
             )